fix(zones): keep the last-owner rule when removing group ownership

+++ lib/Application/Controller/QuickRemoveGroupZoneController.php
@@ -4,7 +4,7 @@
  *  See <https://www.poweradmin.org> for more details.
  *
  *  Copyright 2007-2010 Rejo Zenger <rejo@zenger.nl>
- *  Copyright 2010-2025 Poweradmin Development Team
+ *  Copyright 2010-2026 Poweradmin Development Team
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -25,7 +25,7 @@
  *
  * @package     Poweradmin
  * @copyright   2007-2010 Rejo Zenger <rejo@zenger.nl>
- * @copyright   2010-2025 Poweradmin Development Team
+ * @copyright   2010-2026 Poweradmin Development Team
  * @license     https://opensource.org/licenses/GPL-3.0 GPL
  */
 
@@ -35,6 +35,7 @@ use InvalidArgumentException;
 use Poweradmin\Application\Service\AuditService;
 use Poweradmin\Application\Service\ZoneGroupService;
 use Poweradmin\BaseController;
+use Poweradmin\Domain\Service\ZoneOwnershipModeService;
 
 class QuickRemoveGroupZoneController extends BaseController
 {
@@ -45,8 +46,13 @@ class QuickRemoveGroupZoneController extends BaseController
         parent::__construct($request);
 
         $groupRepository = $this->createUserGroupRepository();
-        $zoneRepository = $this->createZoneGroupRepository();
-        $this->zoneGroupService = new ZoneGroupService($zoneRepository, $groupRepository);
+        $zoneGroupRepository = $this->createZoneGroupRepository();
+        $this->zoneGroupService = new ZoneGroupService(
+            $zoneGroupRepository,
+            $groupRepository,
+            $this->createZoneRepository(),
+            new ZoneOwnershipModeService($this->config)
+        );
     }
 
     public function run(): void
@@ -82,6 +88,13 @@ class QuickRemoveGroupZoneController extends BaseController
             return;
         }
 
+        $refusal = $this->zoneGroupService->getGroupRemovalRefusal($zoneId, $groupId);
+        if ($refusal !== null) {
+            $this->setMessage('edit_group', 'error', $this->refusalMessage($refusal));
+            $this->redirect('/groups/' . $groupId . '/edit');
+            return;
+        }
+
         try {
             $success = $this->zoneGroupService->removeGroupFromZone($zoneId, $groupId);
 
@@ -97,5 +110,27 @@ class QuickRemoveGroupZoneController extends BaseController
         }
 
         $this->redirect('/groups/' . $groupId . '/edit');
+    }
+
+    private function refusalMessage(string $refusal): string
+    {
+        $ownershipMode = new ZoneOwnershipModeService($this->config);
+
+        if ($refusal === ZoneGroupService::REFUSAL_LAST_GROUP_GROUPS_ONLY) {
+            return _('Cannot remove the last group: zone ownership mode is groups_only and requires at least one group. Add another group first.');
+        }
+        if ($refusal === ZoneGroupService::REFUSAL_USERS_ONLY_NO_USER_OWNERS) {
+            return _('Cannot remove group: zone ownership mode is users_only and the zone has no user owners. Add a user owner first.');
+        }
+
+        if ($ownershipMode->isUserOwnerAllowed() && $ownershipMode->isGroupOwnerAllowed()) {
+            $hint = _('Add another owner or a group first.');
+        } elseif ($ownershipMode->isUserOwnerAllowed()) {
+            $hint = _('Add another user owner first (zone ownership mode is users_only).');
+        } else {
+            $hint = _('Add a group first (zone ownership mode is groups_only).');
+        }
+
+        return _('Cannot remove the last owner: this would leave the zone with no ownership.') . ' ' . $hint;
     }
 }
